fix: Order/OrderItem Entity 및 DTO (#7)
- orderPrice 필드 타입 BigDecimal -> Integer 변경
- 그에 따른 Responese도 변경

+++ src/main/java/com/shoppingmall/ecommerceapi/domain/order/dto/AdminOrderItemResponse.java
@@ -16,6 +16,6 @@ public class AdminOrderItemResponse {
   private Long productId;
   private String productName;
   private Integer quantity;
-  private BigDecimal orderPrice;
+  private Integer orderPrice;
   private BigDecimal totalPrice;
 }
