+++ src/main/java/com/shoppingmall/ecommerceapi/domain/order/dto/OrderItemResponse.java
@@ -16,6 +16,6 @@ public class OrderItemResponse {
   private Long productId;
   private String productName;
   private Integer quantity;
-  private BigDecimal orderPrice;
+  private Integer orderPrice;
   private BigDecimal totalPrice;
 }
